requirements.txt
gunicorn
psycopg2-binary
whitenoise
django
djangorestframework
django-cors-headers
